Restructure block validation directory
